feat(gwasUnifiedFlowFileOrganization): Created Steps directory and selectStudyPopulation step
